Refactored fixtures to conftest
